@@ -1,43 +1,51 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development NODE_OPTIONS=--openssl-legacy-provider node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "npm run development -- --watch",
-        "watch-poll": "npm run watch -- --watch-poll",
-        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --disable-host-check --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "development": "mix",
+        "watch": "mix watch",
+        "watch-poll": "mix watch -- --watch-options-poll=1000",
+        "hot": "mix watch --hot",
+        "production": "mix --production"
     },
     "devDependencies": {
+        "@vue/compiler-sfc": "^3.2.31",
         "axios": "^0.21.0",
         "bootstrap": "^4.0.0",
         "cross-env": "^7.0",
+        "css-loader": "^6.6.0",
         "elliptic": "^6.5.4",
+        "file-loader": "^6.2.0",
         "jquery": "^3.2",
         "lang.js": "^1.1.14",
-        "laravel-mix": "^5.0.7",
+        "laravel-mix": "^6.0.42",
         "lodash": "^4.17.20",
+        "mini-css-extract-plugin": "^2.5.3",
         "node-forge": "0.10.0",
         "popper.js": "^1.12",
         "pusher-js": "^7.0.0",
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.20.1",
         "sass-loader": "^8.0.0",
-        "vue": "^2.5.17",
-        "vue-axios": "^2.1.5",
-        "vue-template-compiler": "^2.6.10"
+        "url-loader": "^4.1.1",
+        "vue": "^3.0",
+        "vue-axios": "^3.4.1",
+        "webpack": "^5.69.1",
+        "webpack-cli": "^4.9.2",
+        "webpack-dev-server": "^4.7.4"
     },
     "dependencies": {
+        "@coreui/vue": "^4.1.2",
         "@fortawesome/fontawesome-svg-core": "^1.2.32",
         "@fortawesome/free-brands-svg-icons": "^5.15.1",
         "@fortawesome/free-regular-svg-icons": "^5.15.1",
         "@fortawesome/free-solid-svg-icons": "^5.15.1",
-        "@fortawesome/vue-fontawesome": "^2.0.0",
+        "@fortawesome/vue-fontawesome": "^3.0.0-5",
         "laravel-echo": "^1.9.0",
         "moment": "^2.29.1",
-        "vue-datetime": "^1.0.0-beta.14",
-        "vue-load-image": "^0.1.12",
-        "vue-router": "^3.5.1"
+        "vue-load-image": "^1.1.0",
+        "vue-loader": "^17.0.0",
+        "vue-router": "^4.0.12",
+        "vue3-date-time-picker": "^2.6.0",
+        "vue3-langjs": "^1.0.2"
     }
 }
