@@ -4,18 +4,18 @@
             <h2>
                 <font-awesome-icon icon="spinner" spin class="fa-icon" v-if="loading"/>
                 <font-awesome-icon :icon="['fas', 'chevron-right']" class="fa-icon" v-if="!loading" fixed-width/>
-                {{ 'main.user_settings.personal_information' | trans }}
+                {{ this.$t('main.user_settings.personal_information') }}
             </h2>
         </div>
         <div class="text-left">
             <div class="mt-4 alert alert-success text-center fadeIn" role="alert" v-if="status === STATUS_OK">
-                <h2 class="alert-heading">{{ 'main.information_update_success' | trans }}</h2>
+                <h2 class="alert-heading">{{ this.$t('main.information_update_success') }}</h2>
             </div>
             <div class="mt-4 alert alert-danger fadeIn" role="alert" v-if="status === STATUS_NOT_OK">
-                <h2 class="alert-heading">{{ 'main.information_update_fail' | trans }}</h2>
+                <h2 class="alert-heading">{{ this.$t('main.information_update_fail') }}</h2>
                 <ul class="list">
                     <li v-for="(errorMessages, index) in errors">
-                        <span>{{ 'validation.error_in_input' | trans }} {{ 'validation.attributes.' + index | trans }}</span>
+                        <span>{{ this.$t('validation.error_in_input') }} {{ this.$t('validation.attributes.' + index) }}</span>
                         <ul class="list-unstyled">
                             <li v-for="errorMessage in errorMessages">
                                 {{ errorMessage }}
@@ -33,7 +33,7 @@
                         <div class="form-group row mb-4">
                             <div class="col-md-8">
                                 <label class="col-form-label text-md-right" for="phone">{{
-                                        'main.phone' | trans
+                                        this.$t('main.phone')
                                     }}</label>
                                 <input v-model="user.phone" class="form-control" maxlength="11" type="tel" id="phone"
                                        name="phone" placeholder="370..."
@@ -43,7 +43,7 @@
                         <div class="form-group row">
                             <div class="col-md-6">
                                 <button type="submit" class="btn btn-primary">
-                                    {{ 'main.save' | trans }}
+                                    {{ this.$t('main.save') }}
                                 </button>
                             </div>
                         </div>
